tmp: improve test setup
